\docType{package}
\name{cvGWAS-package}
\alias{cvgwas}
\alias{cvGWAS}
\alias{cvGWAS-package}
\title{GWAS based on predictive ability}
\description{
  cvGWAS: Genome-wide association analysis based on
  predictive ability
}
\details{
  \tabular{ll}{ Package: \tab cvGWAS\cr Type: \tab
  Package\cr Version: \tab 0.1\cr Date: \tab 2013-07-29\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr
  Depends: \tab GenABEL\cr }

  The package provides an analytical tool for genome-wide
  association study based on cross-validated predictive
  ability of each genetic variant.
}
\examples{
\dontrun{
demo(ge03d2)
demo(srdta)
demo(srdtawin)
}
}
\author{
  Xia Shen
}
\references{
  To cite this work:

  Shen X. (2013) The missing heritability revealed in
  \emph{Arabidopsis thaliana}. \bold{Submitted}.
}
\seealso{
  \code{DatABEL}, \code{genetics}, \code{haplo.stats},
  \code{qvalue}
}
\keyword{ability}
\keyword{cross-validation,}
\keyword{GWAS,}
\keyword{predictive}

